fix(graph): move and rename things
